safari-600.7-branch/LayoutTests/fast/regions/region-element-display-restriction-expected.txt

Test that only non-replaced block elements can be transformed into regions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS createRegionElement("div", "none") is false
PASS createRegionElement("div", "block") is true
PASS createRegionElement("div", "inline-block") is true
PASS createRegionElement("div", "compact") is false
PASS createRegionElement("div", "inline") is false
PASS createRegionElement("div", "table") is false
PASS createRegionElement("div", "inline-table") is false
PASS createRegionElement("div", "table-cell") is true
PASS createRegionElement("div", "table-caption") is true
PASS createRegionElement("div", "list-item") is true
PASS createRegionElement("div", "-webkit-box") is false
PASS createRegionElement("div", "-webkit-inline-box") is false
PASS createRegionElement("div", "-webkit-flex") is false
PASS createRegionElement("div", "-webkit-inline-flex") is false
PASS successfullyParsed is true

TEST COMPLETE
